requirements.txt
numpy
opencv-python
python-dotenv
tqdm

-f https://dl.fbaipublicfiles.com/detectron2/wheels/cu113/torch1.10/index.html
detectron2 

-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.0+cu113 
torchvision==0.11.1+cu113